Squash sets into the node hiearchy
There are no more separate set properties (eg for bonds). Instead, a bond is a node with two children and type bond. Nodes can now be the child of more than one parent. An alias is used internally to provide the second parent. This allowed me to delete lots of code and will make other backends much easier.