use boost exceptions to improve error messages
It allows us to add info above the throw point. So we can have the inferface methods (in NodeHandle or decorators) add info that they know about. This reduces the number of places info has to be added and allows us to add some info that is not accessible closer to throw (eg semantics)