Fix and test storage of paths in RMF files.
Paths to external files should always be stored as
paths relative to the directory containing the RMF
file - this helps to ensure the external file can
still be found if the directory structure is archived.
However, the API should accept both relative (to the
current working directory) and absolute paths as input,
and always return absolute paths as output.