Track the starting residue # in structure provenance.
The range of residues read from the PDB file might not match
those in the model (for example, if residues are not numbered
starting at 1, or if two PDB files are concatenated). Record
the starting residue number (essentially equivalent to storing
the offset between the two numbering systems). Note that it isn't
necessary to also store the *ending* residue number - the number
of residues to read can be deduced from those in the model.