Add a better test for OpenMP 3
Our previous test for OpenMP 3 used a pragma
that was only present in OpenMP 3 (not 2).
However, this didn't work because most compilers
simply ignore unknown pragmas. Replace with a call
to a function that's only present in OpenMP 3, so
that compilation will decisively fail. This should
fix the Mac 10.6 build, which was including OpenMP
support even though the gcc version in OS X 10.6
only supports OpenMP 2.5.

As an aside, recent cmake versions extract the OpenMP
vesion number, which would simplify this test in the
future once we can require a new enough cmake.