Search for python2/python3 binaries with old CMake
The old CMake logic only looks for a Python binary
called 'python'. This is no good on RHEL8 since it
has no /usr/bin/python but also does not have a new
enough CMake to use the new CMake logic. Update the
old logic to also look for a "python2" or "python3"
binary.