Warn about missing RMF_OVERRIDE
If we override a virtual method in a subclass, we
should use RMF_OVERRIDE so that the compiler reports
an error if we got the signature wrong (such that it
hides the base method instead of overriding it).
Add -Wsuggest-override to gcc warnings (if gcc is new
enough) to point out places where we missed this.